change data type of price to float

+++ database/migrations/2024_11_14_135414_create_orders_table.php
@@ -23,9 +23,9 @@ return new class extends Migration
             $table->foreignId('cakeId');
             $table->foreign('cakeId')->references('id')->on('cakes')->cascadeOnDelete();
             
-            $table->bigInteger('price')->nullable();
+            $table->float('price')->nullable();
             $table->integer('quantity')->default(1);
-            $table->bigInteger('discount')->nullable();
+            $table->float('discount')->nullable();
 
             $this->getDefaultTimestamps($table);
         });
